+++ tests/integration/distributed-replicas/Dockerfile
