+++ tests/integration/distributed-replicas/config.yml
